requirements.txt
Cython==0.29.21
deepspeed==0.3.13
future==0.18.2
imageio==2.9.0
imgaug==0.4.0
matplotlib==3.2.2
ninja==1.10.0.post2
numpy==1.19.5
opencv-python==4.3.0.36
ordered-set==4.0.2
pandas==1.0.5
Pillow==6.2.2
requests==2.23.0
requests-oauthlib==1.3.0
scikit-image==0.17.2
scikit-learn==0.23.2
six==1.14.0
tensorboard==2.1.1
torch==1.8.0
torchvision==0.9.0
tqdm==4.43.0
urllib3==1.25.8
